+++ R/org-process.R
@@ -19,7 +19,7 @@ is_org_process <- function(file,
                            control = FALSE
                            ) {
   GEO <- NULL
-
+  is_debug(deep = TRUE)
   dots <- get_innlesarg(spec = filespec)
 
   ## For GEO codes that are derived from a combination of two columns