+++ R/utils-misc.R
@@ -26,7 +26,8 @@ reset_opt <- reset_options
 #' @param opt Debug options
 #' @param val Value options
 #' @details Options available:
-#'   - "fun"       : Activating options `orgdata.debug`
+#'   - "shallow"   : Show the executed funcitons in the process flow
+#'   - "deep"      : Show deeper including helper functions in the process flow
 #'   - "nrow"      : For `orgdata.debug.nrow`. Default value is 20 if `val` argument is missing
 #'   - "row"       : For `orgdata.debug.row`. Default value is 1:50 if `val` argument is missing
 #'   - "aggregate" : Activating `orgdata.debug.aggregate`
@@ -37,10 +38,11 @@ reset_opt <- reset_options
 #' debug_opt("nrow", 30)
 #' }
 #' @export
-debug_opt <- function(opt = c("fun", "nrow", "row", "aggregate", "geo"), val = NULL){
+debug_opt <- function(opt = c("shallow", "deep", "nrow", "row", "aggregate", "geo"), val = NULL){
   opt <- match.arg(opt)
   switch(opt,
-         fun = options(orgdata.debug = TRUE),
+         shallow = options(orgdata.debug = TRUE),
+         deep = options(orgdata.debug = "deep"),
          nrow = options(orgdata.debug.nrow = ifelse(is.null(val), 20, val)),
          row = options(orgdata.debug.row = ifelse(is.null(val), 1:50, val)),
          aggregate = options(orgdata.debug.aggregate = TRUE),